requirements.txt
fastapi
uvicorn
line-bot-sdk
gspread
google-auth
anthropic